# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopping site API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
}

export interface LoginResponse {
  user: User;
  message: string;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  description: string;
  imageUrl: string;
  rating: number;
  category: string;
}

export interface CartItem {
  productId: number;
  product: Product;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
}

export interface AddToCartBody {
  productId: number;
  quantity?: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export interface CheckoutBody {
  bankName: string;
  accountNumber: string;
  accountHolder: string;
}

export interface Account {
  id: number;
  username: string;
  name?: string;
  email?: string;
  phone?: string;
  age?: number;
}

export interface UpdateAccountBody {
  name?: string;
  email?: string;
  phone?: string;
  age?: number;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetProductsParams = {
  search?: string;
};