# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopping site API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with username and password
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
  }),
  message: zod.string(),
});

/**
 * @summary Logout current session
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
});

/**
 * @summary Get all products with optional search
 */
export const GetProductsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
});

export const GetProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  price: zod.number(),
  description: zod.string(),
  imageUrl: zod.string(),
  rating: zod.number(),
  category: zod.string(),
});
export const GetProductsResponse = zod.array(GetProductsResponseItem);

/**
 * @summary Get current user's cart
 */
export const GetCartResponse = zod.object({
  items: zod.array(
    zod.object({
      productId: zod.number(),
      product: zod.object({
        id: zod.number(),
        name: zod.string(),
        price: zod.number(),
        description: zod.string(),
        imageUrl: zod.string(),
        rating: zod.number(),
        category: zod.string(),
      }),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Add product to cart
 */
export const AddToCartBody = zod.object({
  productId: zod.number(),
  quantity: zod.number().optional(),
});

export const AddToCartResponse = zod.object({
  items: zod.array(
    zod.object({
      productId: zod.number(),
      product: zod.object({
        id: zod.number(),
        name: zod.string(),
        price: zod.number(),
        description: zod.string(),
        imageUrl: zod.string(),
        rating: zod.number(),
        category: zod.string(),
      }),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Update cart item quantity
 */
export const UpdateCartItemParams = zod.object({
  productId: zod.coerce.number(),
});

export const UpdateCartItemBody = zod.object({
  quantity: zod.number(),
});

export const UpdateCartItemResponse = zod.object({
  items: zod.array(
    zod.object({
      productId: zod.number(),
      product: zod.object({
        id: zod.number(),
        name: zod.string(),
        price: zod.number(),
        description: zod.string(),
        imageUrl: zod.string(),
        rating: zod.number(),
        category: zod.string(),
      }),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Remove item from cart
 */
export const RemoveFromCartParams = zod.object({
  productId: zod.coerce.number(),
});

export const RemoveFromCartResponse = zod.object({
  items: zod.array(
    zod.object({
      productId: zod.number(),
      product: zod.object({
        id: zod.number(),
        name: zod.string(),
        price: zod.number(),
        description: zod.string(),
        imageUrl: zod.string(),
        rating: zod.number(),
        category: zod.string(),
      }),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Checkout and clear cart
 */
export const CheckoutBody = zod.object({
  bankName: zod.string(),
  accountNumber: zod.string(),
  accountHolder: zod.string(),
});

export const CheckoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get account details
 */
export const GetAccountResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string().optional(),
  email: zod.string().optional(),
  phone: zod.string().optional(),
  age: zod.number().optional(),
});

/**
 * @summary Update account details
 */
export const UpdateAccountBody = zod.object({
  name: zod.string().optional(),
  email: zod.string().optional(),
  phone: zod.string().optional(),
  age: zod.number().optional(),
});

export const UpdateAccountResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string().optional(),
  email: zod.string().optional(),
  phone: zod.string().optional(),
  age: zod.number().optional(),
});