# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopping site API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Account {
  id: number;
  username: string;
  name?: string;
  email?: string;
  phone?: string;
  age?: number;
}