# lib/api-zod/src/generated/types/updateAccountBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopping site API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAccountBody {
  name?: string;
  email?: string;
  phone?: string;
  age?: number;
}